♻️ refactor: 정렬&필터 메모이제이션
