+++ src/app/(default)/(ticket)/ticket-result/Filter.tsx
@@ -5,10 +5,15 @@ import Accordion from "@/components/Accordion/Accordion";
 import AccordionBody from "@/components/Accordion/AccordionBody";
 import AccordionHeader from "@/components/Accordion/AccordionHeader";
 import AccordionItem from "@/components/Accordion/AccordionItem";
-import Badge from "@/components/Badge/Badge";
-import useAllChecked from "@/hook/useAllChecked";
+import AirlineFilter from "@/components/Filters/AirlineFilter";
+import NonStopFilter from "@/components/Filters/NonstopFilter";
+import OriginArrTime from "@/components/Filters/OriginArrTime";
+import OriginDepTime from "@/components/Filters/OriginDepTime";
+import PriceFilter from "@/components/Filters/PriceFilter";
+import ReturnArrTime from "@/components/Filters/ReturnArrTime";
+import ReturnDepTime from "@/components/Filters/ReturnDepTime";
 import { AirlineData, CodeState } from "@/types";
-import { ChangeEvent, memo, useEffect, useRef, useState } from "react";
+import { memo } from "react";
 import "./Filter.scss";
 
 const Filter = memo(
@@ -27,249 +32,7 @@ const Filter = memo(
     prices: number[];
     handleFilterChange: (filter: FilterProps) => void;
   }) => {
-    const [isNonStop, setIsNonStop] = useState(nonStop);
-    const [originDepTime, setOriginDepTime] = useState<number[]>([
-      6, 12, 18, 24,
-    ]);
-    const [originArrTime, setOriginArrTime] = useState<number[]>([
-      6, 12, 18, 24,
-    ]);
-    const [returnDepTime, setReturnDepTime] = useState<number[]>([
-      6, 12, 18, 24,
-    ]);
-    const [returnArrTime, setReturnArrTime] = useState<number[]>([
-      6, 12, 18, 24,
-    ]);
-    const [maxPrice, setMaxPrice] = useState(Math.max(...prices));
-    const [selectedAirlines, setSelectedAirlines] =
-      useState<string[]>(carrierCodes);
-    const airlineRef = useRef(null);
-
-    console.log("필터 렌더링");
-
-    /* -------------------------------------------------------------------------- */
-    /*                          경유 선택 시 직항 경유 변경 처리                         */
-    /* -------------------------------------------------------------------------- */
-    const handleNonStopChange = () => {
-      setIsNonStop(!isNonStop);
-      handleFilterChange({ nonStop: !isNonStop });
-    };
-
-    /* -------------------------------------------------------------------------- */
-    /*                             출도착 시간 변경 처리                               */
-    /* -------------------------------------------------------------------------- */
-    useEffect(() => {
-      console.log("출발 출발시간 변경");
-      handleFilterChange({ originDepTime });
-    }, [originDepTime]);
-
-    const handleOriginDepChange = (e: ChangeEvent<HTMLInputElement>) => {
-      const value = Number(e.target.value);
-
-      setOriginDepTime((prev) => {
-        if (!prev) {
-          return [value];
-        }
-
-        if (e.target.checked) {
-          return [...prev, value];
-        }
-        return prev.filter((time) => time !== value);
-      });
-    };
-
-    useEffect(() => {
-      console.log("도착 출발시간 변경");
-      handleFilterChange({ returnDepTime });
-    }, [returnDepTime]);
-
-    const handleReturnDepChange = (e: ChangeEvent<HTMLInputElement>) => {
-      const value = Number(e.target.value);
-
-      setReturnDepTime((prev) => {
-        if (!prev) {
-          return [value];
-        }
-
-        if (e.target.checked) {
-          return [...prev, value];
-        }
-        return prev.filter((time) => time !== value);
-      });
-    };
-
-    useEffect(() => {
-      console.log("출발 도착시간 변경");
-      handleFilterChange({ originArrTime });
-    }, [originArrTime]);
-
-    const handleOriginArrChange = (e: ChangeEvent<HTMLInputElement>) => {
-      const value = Number(e.target.value);
-
-      setOriginArrTime((prev) => {
-        if (!prev) {
-          return [value];
-        }
-
-        if (e.target.checked) {
-          return [...prev, value];
-        }
-        return prev.filter((time) => time !== value);
-      });
-    };
-
-    useEffect(() => {
-      console.log("도착 도착시간 변경");
-      handleFilterChange({ returnArrTime });
-    }, [returnArrTime]);
-
-    const handleReturnArrChange = (e: ChangeEvent<HTMLInputElement>) => {
-      const value = Number(e.target.value);
-
-      setReturnArrTime((prev) => {
-        if (!prev) {
-          return [value];
-        }
-
-        if (e.target.checked) {
-          return [...prev, value];
-        }
-        return prev.filter((time) => time !== value);
-      });
-    };
-
-    /* -------------------------------------------------------------------------- */
-    /*                                   가격 설정                                  */
-    /* -------------------------------------------------------------------------- */
-    const handlePriceChange = (e: ChangeEvent<HTMLInputElement>) => {
-      const { value, style, max } = e.target;
-      const newMaxPrice = Number(value);
-      setMaxPrice(newMaxPrice);
-      handleFilterChange({ maxPrice: newMaxPrice });
-      const percent = 100 / +max;
-      style.background = `linear-gradient(to right, var(--color-primary) 0%, var(--color-primary) ${percent * +value}%, var(--color-gray-50) ${percent * +value}%, var(--color-gray-50) 100%`;
-    };
-    /* -------------------------------------------------------------------------- */
-    /*                                 항공사 선택                                   */
-    /* -------------------------------------------------------------------------- */
-    const allianceCont = [
-      {
-        name: "Skyteam",
-        title: "스카이팀",
-        content: "",
-        checked: true,
-      },
-      {
-        name: "Star Alliance",
-        title: "스타얼라이언스",
-        content: "",
-        checked: true,
-      },
-      {
-        name: "oneworld",
-        title: "원월드",
-        content: "",
-        checked: true,
-      },
-      {
-        name: "others",
-        title: "기타",
-        content: "",
-        checked: true,
-      },
-    ];
-
-    const [allianceCheck, setAllianceCheck] = useState(allianceCont);
-    const { setCheck } = useAllChecked(allianceCheck, setAllianceCheck);
-
-    const handleSelectAll = () => {
-      setAllianceCheck(
-        allianceCheck.map((item) => ({ ...item, checked: true })),
-      );
-    };
-
-    const handleUnselectAll = () => {
-      setAllianceCheck(
-        allianceCheck.map((item) => ({ ...item, checked: false })),
-      );
-    };
-
-    /* -------------------------------------------------------------------------- */
-    /*                                  동맹체 체크 여부                              */
-    /* -------------------------------------------------------------------------- */
-    useEffect(() => {
-      const updatedAirlines: string[] = [];
-
-      allianceCheck.forEach((alliance) => {
-        if (alliance.checked) {
-          carrierCodes.forEach((carrierCode) => {
-            if (airline[carrierCode].allianceEng === alliance.name) {
-              updatedAirlines.push(carrierCode);
-            } else if (
-              airline[carrierCode].allianceEng === "" &&
-              alliance.name === "others"
-            ) {
-              updatedAirlines.push(carrierCode);
-            }
-          });
-        }
-      });
-
-      setSelectedAirlines(updatedAirlines);
-    }, [allianceCheck]);
-
-    const handleAllianceChk = (e: ChangeEvent<HTMLInputElement>) => {
-      setCheck(e);
-    };
-
-    const alliances = allianceCheck.map((item) => (
-      <li key={item.title}>
-        <input
-          type="checkbox"
-          name={item.name}
-          id={item.name}
-          value={item.title}
-          checked={item.checked}
-          onChange={handleAllianceChk}
-        />
-        <label htmlFor={item.name}>{item.title}</label>
-      </li>
-    ));
-
-    useEffect(() => {
-      handleFilterChange({ airline: selectedAirlines });
-    }, [selectedAirlines]);
-
-    const handleAirlineChk = (e: ChangeEvent<HTMLInputElement>) => {
-      const { value } = e.target;
-
-      setSelectedAirlines((prev) => {
-        if (!prev) {
-          return [value];
-        }
-
-        if (e.target.checked) {
-          return [...prev, value];
-        }
-        return prev.filter((time) => time !== value);
-      });
-    };
-
-    const airlines = carrierCodes.map((item, idx) => {
-      return (
-        <li key={idx}>
-          <input
-            ref={airlineRef}
-            type="checkbox"
-            id={item}
-            value={item}
-            checked={selectedAirlines.includes(item)}
-            onChange={handleAirlineChk}
-          />
-          <label htmlFor={item}>{airline[item].nameKor}</label>
-        </li>
-      );
-    });
+    console.log("Filter 렌더링");
 
     return (
       <div className="filter">
@@ -281,18 +44,10 @@ const Filter = memo(
                 <p className="filter-title">경유</p>
               </AccordionHeader>
               <AccordionBody>
-                <div className="filter-contents nonstop">
-                  <input
-                    type="checkbox"
-                    id="nonstop-filter"
-                    name="nonstop"
-                    checked={isNonStop}
-                    onChange={handleNonStopChange}
-                  />
-                  <label htmlFor="nonstop-filter">
-                    <Badge selected={isNonStop}>직항</Badge>
-                  </label>
-                </div>
+                <NonStopFilter
+                  nonStop={nonStop}
+                  handleFilterChange={handleFilterChange}
+                />
               </AccordionBody>
             </AccordionItem>
           </Accordion>
@@ -305,123 +60,9 @@ const Filter = memo(
               <p className="filter-title">출발시간</p>
             </AccordionHeader>
             <AccordionBody>
-              <div>
-                <p className="filter-subtitle">가는 날</p>
-                <ul className="filter-contents time">
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-depTime-0006"
-                      value={6}
-                      onChange={handleOriginDepChange}
-                      checked={originDepTime.includes(6)}
-                    />
-                    <label htmlFor="origin-depTime-0006">
-                      <span />
-                    </label>
-                    <p className="time-legend first">0시</p>
-                  </li>
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-depTime-0612"
-                      value={12}
-                      onChange={handleOriginDepChange}
-                      checked={originDepTime.includes(12)}
-                    />
-                    <label htmlFor="origin-depTime-0612">
-                      <span />
-                    </label>
-                    <p className="time-legend">6시</p>
-                  </li>
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-depTime-1218"
-                      value={18}
-                      onChange={handleOriginDepChange}
-                      checked={originDepTime.includes(18)}
-                    />
-                    <label htmlFor="origin-depTime-1218">
-                      <span />
-                    </label>
-                    <p className="time-legend">12시</p>
-                  </li>
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-depTime-1824"
-                      value={24}
-                      onChange={handleOriginDepChange}
-                      checked={originDepTime.includes(24)}
-                    />
-                    <label htmlFor="origin-depTime-1824">
-                      <span />
-                    </label>
-                    <p className="time-legend">18시</p>
-                    <p className="time-legend last">24시</p>
-                  </li>
-                </ul>
-              </div>
+              <OriginDepTime handleFilterChange={handleFilterChange} />
               {tripType === "round" && (
-                <div>
-                  <p className="filter-subtitle">오는 날</p>
-                  <ul className="filter-contents time">
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-depTime-0006"
-                        value={6}
-                        checked={returnDepTime.includes(6)}
-                        onChange={handleReturnDepChange}
-                      />
-                      <label htmlFor="return-depTime-0006">
-                        <span />
-                        <p className="time-legend first">0시</p>
-                      </label>
-                    </li>
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-depTime-0612"
-                        value={12}
-                        checked={returnDepTime.includes(12)}
-                        onChange={handleReturnDepChange}
-                      />
-                      <label htmlFor="return-depTime-0612">
-                        <span />
-                        <p className="time-legend">6시</p>
-                      </label>
-                    </li>
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-depTime-1218"
-                        value={18}
-                        checked={returnDepTime.includes(18)}
-                        onChange={handleReturnDepChange}
-                      />
-                      <label htmlFor="return-depTime-1218">
-                        <span />
-                        <p className="time-legend">12시</p>
-                      </label>
-                    </li>
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-depTime-1824"
-                        value={24}
-                        checked={returnDepTime.includes(24)}
-                        onChange={handleReturnDepChange}
-                      />
-                      <label htmlFor="return-depTime-1824">
-                        <span />
-                        <p className="time-legend">18시</p>
-                        <p className="time-legend last">24시</p>
-                      </label>
-                    </li>
-                  </ul>
-                </div>
+                <ReturnDepTime handleFilterChange={handleFilterChange} />
               )}
             </AccordionBody>
           </AccordionItem>
@@ -434,123 +75,9 @@ const Filter = memo(
               <p className="filter-title">도착시간</p>
             </AccordionHeader>
             <AccordionBody>
-              <div>
-                <p className="filter-subtitle">가는 날</p>
-                <ul className="filter-contents time">
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-arrTime-0006"
-                      value={6}
-                      checked={originArrTime.includes(6)}
-                      onChange={handleOriginArrChange}
-                    />
-                    <label htmlFor="origin-arrTime-0006">
-                      <span />
-                      <p className="time-legend first">0시</p>
-                    </label>
-                  </li>
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-arrTime-0612"
-                      value={12}
-                      checked={originArrTime.includes(12)}
-                      onChange={handleOriginArrChange}
-                    />
-                    <label htmlFor="origin-arrTime-0612">
-                      <span />
-                      <p className="time-legend">6시</p>
-                    </label>
-                  </li>
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-arrTime-1218"
-                      value={18}
-                      checked={originArrTime.includes(18)}
-                      onChange={handleOriginArrChange}
-                    />
-                    <label htmlFor="origin-arrTime-1218">
-                      <span />
-                      <p className="time-legend">12시</p>
-                    </label>
-                  </li>
-                  <li className="time-selection">
-                    <input
-                      type="checkbox"
-                      id="origin-arrTime-1824"
-                      value={24}
-                      checked={originArrTime.includes(24)}
-                      onChange={handleOriginArrChange}
-                    />
-                    <label htmlFor="origin-arrTime-1824">
-                      <span />
-                      <p className="time-legend">18시</p>
-                      <p className="time-legend last">24시</p>
-                    </label>
-                  </li>
-                </ul>
-              </div>
+              <OriginArrTime handleFilterChange={handleFilterChange} />
               {tripType === "round" && (
-                <div>
-                  <p className="filter-subtitle">오는 날</p>
-                  <ul className="filter-contents time">
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-arrTime-0006"
-                        value={6}
-                        checked={returnArrTime.includes(6)}
-                        onChange={handleReturnArrChange}
-                      />
-                      <label htmlFor="return-arrTime-0006">
-                        <span />
-                        <p className="time-legend first">0시</p>
-                      </label>
-                    </li>
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-arrTime-0612"
-                        value={12}
-                        checked={returnArrTime.includes(12)}
-                        onChange={handleReturnArrChange}
-                      />
-                      <label htmlFor="return-arrTime-0612">
-                        <span />
-                        <p className="time-legend">6시</p>
-                      </label>
-                    </li>
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-arrTime-1218"
-                        value={18}
-                        checked={returnArrTime.includes(18)}
-                        onChange={handleReturnArrChange}
-                      />
-                      <label htmlFor="return-arrTime-1218">
-                        <span />
-                        <p className="time-legend">12시</p>
-                      </label>
-                    </li>
-                    <li className="time-selection">
-                      <input
-                        type="checkbox"
-                        id="return-arrTime-1824"
-                        value={24}
-                        checked={returnArrTime.includes(24)}
-                        onChange={handleReturnArrChange}
-                      />
-                      <label htmlFor="return-arrTime-1824">
-                        <span />
-                        <p className="time-legend">18시</p>
-                        <p className="time-legend last">24시</p>
-                      </label>
-                    </li>
-                  </ul>
-                </div>
+                <ReturnArrTime handleFilterChange={handleFilterChange} />
               )}
             </AccordionBody>
           </AccordionItem>
@@ -563,48 +90,11 @@ const Filter = memo(
               <p className="filter-title">항공사 선택</p>
             </AccordionHeader>
             <AccordionBody>
-              <div className="filter-contents airlines">
-                <ul className="all-check">
-                  <li>
-                    <input
-                      type="checkbox"
-                      id="allChk"
-                      name="allChk"
-                      value="allChk"
-                      disabled={
-                        allianceCheck.filter((item) => item.checked !== true)
-                          .length < 1
-                      }
-                      checked={
-                        allianceCheck.filter((item) => item.checked !== true)
-                          .length >= 1
-                      }
-                      onChange={handleSelectAll}
-                    />
-                    <label htmlFor="allChk">모두 선택</label>
-                  </li>
-                  <li>
-                    <input
-                      type="checkbox"
-                      id="allCancel"
-                      name="allChk"
-                      value="allCancel"
-                      disabled={
-                        allianceCheck.filter((item) => item.checked !== true)
-                          .length >= 1
-                      }
-                      checked={
-                        allianceCheck.filter((item) => item.checked !== true)
-                          .length < 1
-                      }
-                      onChange={handleUnselectAll}
-                    />
-                    <label htmlFor="allCancel">모두 해제</label>
-                  </li>
-                </ul>
-                <ul className="alliance">{alliances}</ul>
-                <ul className="airline">{airlines}</ul>
-              </div>
+              <AirlineFilter
+                airline={airline}
+                carrierCodes={carrierCodes}
+                handleFilterChange={handleFilterChange}
+              />
             </AccordionBody>
           </AccordionItem>
         </Accordion>
@@ -616,21 +106,10 @@ const Filter = memo(
               <p className="filter-title">가격대 (탑승객 전체)</p>
             </AccordionHeader>
             <AccordionBody>
-              <div className="filter-contents maxPrice">
-                <label htmlFor="max-price">
-                  {`${maxPrice?.toLocaleString()}원 미만`}
-                </label>
-                <input
-                  type="range"
-                  className="input-range horizontal"
-                  id="max-price"
-                  min={0}
-                  max={Math.max(...prices)}
-                  value={maxPrice}
-                  step={100}
-                  onChange={handlePriceChange}
-                />
-              </div>
+              <PriceFilter
+                prices={prices}
+                handleFilterChange={handleFilterChange}
+              />
             </AccordionBody>
           </AccordionItem>
         </Accordion>
